Fix Date Format in Charts

+++ components/v1/MarketDetailPage/LeveragedTokenChart.tsx
@@ -116,7 +116,7 @@ const LeveragedTokenChart: FunctionComponent<LeveragedTokenChartProps> = ({ addr
                                         const selectedData = payload[0].payload;
                                         const timestamp = selectedData.timestamp;
                                         const date = new Date(timestamp);
-                                        const formattedDate = new Intl.DateTimeFormat("en-US", { hour: "numeric", day: "numeric", month: "numeric", year: "numeric", minute: "numeric" }).format(date);
+                                        const formattedDate = new Intl.DateTimeFormat("en-US", { hour: "numeric", day: "numeric", month: "short", year: "numeric", minute: "numeric"}).format(date);
 
                                         setNAV(selectedData.nav);
                                         const change = ((selectedData.nav - currentData.oldestNAV) / currentData.oldestNAV) * 100;
