+++ components/v1/MarketDetailPage/VaultChart.tsx
@@ -92,7 +92,7 @@ const VaultChart: FunctionComponent<VaultChartProps> = ({ address }) => {
                                         const selectedData = payload[0].payload;
                                         const timestamp = selectedData.timestamp;
                                         const date = new Date(timestamp);
-                                        const formattedDate = new Intl.DateTimeFormat("en-US", { hour: "numeric", day: "numeric", month: "numeric", year: "numeric", minute: "numeric" }).format(date);
+                                        const formattedDate = new Intl.DateTimeFormat("en-US", { hour: "numeric", day: "numeric", month: "short", year: "numeric", minute: "numeric" }).format(date);
 
                                         setSupplyAPY(selectedData.supply_apy);
                                         setBorrowAPY(selectedData.borrow_apy);
